% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{autoPredictor}
\alias{autoPredictor}
\title{Automatically predict freshness profile by using models based on non-invasive sensors.}
\usage{
autoPredictor(
  databaseName,
  productName,
  inputSensor,
  modelType = "regression",
  type = "list"
)
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{productName}{String - a name product being assessed.}

\item{inputSensor}{String - a filename of input sensor data used for the prediction.}

\item{modelType}{String - a type of ML models. "regression" or "classification", default "regression"}

\item{type}{String - data type of result prediction. "list" or "json", default "list".}
}
\description{
Description here.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
