% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{daemon}
\alias{daemon}
\title{Activate daemon to open a port for REST API.}
\usage{
daemon()
}
\description{
Description here.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
