% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{DaemonFunctionManager}
\alias{DaemonFunctionManager}
\title{R6 class which contains all the available functions for prediction.}
\description{
R6 class which contains all the available functions for prediction.

R6 class which contains all the available functions for prediction.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{util}}{Class - class of DaemonUtility.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DaemonFunctionManager-new}{\code{DaemonFunctionManager$new()}}
\item \href{#method-DaemonFunctionManager-clone}{\code{DaemonFunctionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DaemonFunctionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-DaemonFunctionManager-new}{}}}
\subsection{Method \code{new()}}{
Inilialise DaemonUtility - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DaemonFunctionManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DaemonFunctionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DaemonFunctionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DaemonFunctionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
