% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daemon.R
\name{DaemonUtility}
\alias{DaemonUtility}
\title{R6 class as utility of Daemon which is available for all the class, members and functions.}
\description{
R6 class as utility of Daemon which is available for all the class, members and functions.

R6 class as utility of Daemon which is available for all the class, members and functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DaemonUtility-new}{\code{DaemonUtility$new()}}
\item \href{#method-DaemonUtility-errorBomb}{\code{DaemonUtility$errorBomb()}}
\item \href{#method-DaemonUtility-clone}{\code{DaemonUtility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DaemonUtility-new"></a>}}
\if{latex}{\out{\hypertarget{method-DaemonUtility-new}{}}}
\subsection{Method \code{new()}}{
Inilialise DaemonUtility - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DaemonUtility$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DaemonUtility-errorBomb"></a>}}
\if{latex}{\out{\hypertarget{method-DaemonUtility-errorBomb}{}}}
\subsection{Method \code{errorBomb()}}{
BOOOOOOOOOM !!! !!! !!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DaemonUtility$errorBomb(errorContent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorContent}}{String - detailed error message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DaemonUtility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DaemonUtility-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DaemonUtility$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
