% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBuilder.R
\name{MlModelBuilder}
\alias{MlModelBuilder}
\title{Child R6 class to build machine learning models and evaluate their performances.}
\description{
Child R6 class to build machine learning models and evaluate their performances.

Child R6 class to build machine learning models and evaluate their performances.
}
\section{Super class}{
\code{\link[foodguardRanker:ModelBuilderInputDataManager]{foodguardRanker::ModelBuilderInputDataManager}} -> \code{MlModelBuilder}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{experimentId}}{Comment here.}

\item{\code{fullExperimentId}}{Comment here.}

\item{\code{outputDir}}{Comment here.}

\item{\code{analyticalMetaTypes}}{Comment here.}

\item{\code{mergedDataset}}{Comment here.}

\item{\code{resultTable}}{Comment here.}

\item{\code{allPerformances}}{Comment here.}

\item{\code{resultStatistics}}{Comment here.}

\item{\code{util}}{Comment here.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MlModelBuilder-new}{\code{MlModelBuilder$new()}}
\item \href{#method-MlModelBuilder-getExperimentIds}{\code{MlModelBuilder$getExperimentIds()}}
\item \href{#method-MlModelBuilder-getAnalyticalDataIds}{\code{MlModelBuilder$getAnalyticalDataIds()}}
\item \href{#method-MlModelBuilder-createOutputDirs}{\code{MlModelBuilder$createOutputDirs()}}
\item \href{#method-MlModelBuilder-createOverView}{\code{MlModelBuilder$createOverView()}}
\item \href{#method-MlModelBuilder-doPCAs}{\code{MlModelBuilder$doPCAs()}}
\item \href{#method-MlModelBuilder-getAnalyticalMetaTypes}{\code{MlModelBuilder$getAnalyticalMetaTypes()}}
\item \href{#method-MlModelBuilder-doModellings}{\code{MlModelBuilder$doModellings()}}
\item \href{#method-MlModelBuilder-clone}{\code{MlModelBuilder$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="foodguardRanker" data-topic="ModelBuilderInputDataManager" data-id="checkDatasets"><a href='../../foodguardRanker/html/ModelBuilderInputDataManager.html#method-ModelBuilderInputDataManager-checkDatasets'><code>foodguardRanker::ModelBuilderInputDataManager$checkDatasets()</code></a></span></li>
<li><span class="pkg-link" data-pkg="foodguardRanker" data-topic="ModelBuilderInputDataManager" data-id="purifyDatasets"><a href='../../foodguardRanker/html/ModelBuilderInputDataManager.html#method-ModelBuilderInputDataManager-purifyDatasets'><code>foodguardRanker::ModelBuilderInputDataManager$purifyDatasets()</code></a></span></li>
<li><span class="pkg-link" data-pkg="foodguardRanker" data-topic="ModelBuilderInputDataManager" data-id="readDatasets"><a href='../../foodguardRanker/html/ModelBuilderInputDataManager.html#method-ModelBuilderInputDataManager-readDatasets'><code>foodguardRanker::ModelBuilderInputDataManager$readDatasets()</code></a></span></li>
<li><span class="pkg-link" data-pkg="foodguardRanker" data-topic="ModelBuilderInputDataManager" data-id="showDatasetOverviews"><a href='../../foodguardRanker/html/ModelBuilderInputDataManager.html#method-ModelBuilderInputDataManager-showDatasetOverviews'><code>foodguardRanker::ModelBuilderInputDataManager$showDatasetOverviews()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-new"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-new}{}}}
\subsection{Method \code{new()}}{
Initialise MlModelBuilder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$new(experimentInfo, mlRankingInfo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experimentInfo}}{Comment here.}

\item{\code{mlRankingInfo}}{Comment here.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-getExperimentIds"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-getExperimentIds}{}}}
\subsection{Method \code{getExperimentIds()}}{
Get experiment IDs by converting random values into SHA256
Check generateExperimentIds() function at ModelBuilderUtility class
for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$getExperimentIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-getAnalyticalDataIds"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-getAnalyticalDataIds}{}}}
\subsection{Method \code{getAnalyticalDataIds()}}{
Create analytical data IDs - sensor's colnames are combined
and changed into SHA 256. Thise IDs are used for automated
prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$getAnalyticalDataIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-createOutputDirs"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-createOutputDirs}{}}}
\subsection{Method \code{createOutputDirs()}}{
Create output dirs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$createOutputDirs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-createOverView"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-createOverView}{}}}
\subsection{Method \code{createOverView()}}{
Create overview of the experiment in JSON format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$createOverView()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-doPCAs"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-doPCAs}{}}}
\subsection{Method \code{doPCAs()}}{
Do PCA - output several PC scores (e.g., 1-4 scores) and their
meta data sets in CSV format, plus score variance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$doPCAs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-getAnalyticalMetaTypes"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-getAnalyticalMetaTypes}{}}}
\subsection{Method \code{getAnalyticalMetaTypes()}}{
Get types of analytical/meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$getAnalyticalMetaTypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-doModellings"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-doModellings}{}}}
\subsection{Method \code{doModellings()}}{
Do modellings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$doModellings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MlModelBuilder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MlModelBuilder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MlModelBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
