% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{mlRankingsViewer}
\alias{mlRankingsViewer}
\title{Show rankings of ML method in a specific experiment.}
\usage{
mlRankingsViewer(databaseName, id, type = "list")
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{id}{String - an ID which store models built by foodguardRanker::modelBuilder().}

\item{type}{String - data type of result experiment overviews. "list" or "json", default "list".}
}
\description{
Show the whole information of the product's ML rankings - it gets the content of file
'mlRankings.csv' by given ID and show them.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
