% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBuilder.R
\name{modelBuilder}
\alias{modelBuilder}
\title{Build ML models and rank these perforamces for food quality assessments based on non-invasive sensors.}
\usage{
modelBuilder(configFileName = "ModelBuilderConfig")
}
\arguments{
\item{configFileName}{A string representing a name of input config JSON file, default "ModelBuilderConfig".}
}
\description{
foodguardRanker::modelBuilder builds the ML (machine learning) models over all the combinations of the non-invasive sensors
(hereinafter, analytical data), food quality indicators (hereinafter, metadata) and ML methods. Then it compares these
performances with several statistics evaluations (e.g., accuracy for classification and RMSE for regression) as rankings.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
