% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBuilder.R
\name{ModelBuilderInputDataManager}
\alias{ModelBuilderInputDataManager}
\title{R6 class to read, purify and store input experimental datasets.}
\description{
R6 class to read, purify and store input experimental datasets.

R6 class to read, purify and store input experimental datasets.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{experimentInfo}}{List - experimental information extracted by ModelBuilderConfigParser$initialize.}

\item{\code{mlRankingInfo}}{List - ML rankings information extracted by ModelBuilderConfigParser$initialize.}

\item{\code{analyticalDatasets}}{List - analytical data contents linked with their data type (e.g., self$analyticalDatasets$NIR).}

\item{\code{metaDatasets}}{List - meta data contents linked with their data type (e.g., self$metaDatasets$TVC).}

\item{\code{util}}{Class - class of ModelBuilderUtility.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelBuilderInputDataManager-new}{\code{ModelBuilderInputDataManager$new()}}
\item \href{#method-ModelBuilderInputDataManager-readDatasets}{\code{ModelBuilderInputDataManager$readDatasets()}}
\item \href{#method-ModelBuilderInputDataManager-checkDatasets}{\code{ModelBuilderInputDataManager$checkDatasets()}}
\item \href{#method-ModelBuilderInputDataManager-purifyDatasets}{\code{ModelBuilderInputDataManager$purifyDatasets()}}
\item \href{#method-ModelBuilderInputDataManager-showDatasetOverviews}{\code{ModelBuilderInputDataManager$showDatasetOverviews()}}
\item \href{#method-ModelBuilderInputDataManager-clone}{\code{ModelBuilderInputDataManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderInputDataManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderInputDataManager-new}{}}}
\subsection{Method \code{new()}}{
Initialise ModelBuilderInputDataManager - it only activates ModelBuilderUtility class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderInputDataManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderInputDataManager-readDatasets"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderInputDataManager-readDatasets}{}}}
\subsection{Method \code{readDatasets()}}{
Read and extract analytical and meta data files in the input config
file - extracted data contents are stored in self$analyticalDatasets
and self$metaDatasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderInputDataManager$readDatasets(experimentInfo, mlRankingInfo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experimentInfo}}{List - experimental information extracted by ModelBuilderConfigParser$initialize.}

\item{\code{mlRankingInfo}}{List - ML rankings information extracted by ModelBuilderConfigParser$initialize.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderInputDataManager-checkDatasets"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderInputDataManager-checkDatasets}{}}}
\subsection{Method \code{checkDatasets()}}{
Check if extracted analytical and meta dataset are in valid formats
to proceed to PCAa and modellings - actually they are checked by
private$checkAnalyticalDataset and private$checkMetaDataset members.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderInputDataManager$checkDatasets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderInputDataManager-purifyDatasets"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderInputDataManager-purifyDatasets}{}}}
\subsection{Method \code{purifyDatasets()}}{
Purify datasets - set row names, remove NA values and omitted samples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderInputDataManager$purifyDatasets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderInputDataManager-showDatasetOverviews"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderInputDataManager-showDatasetOverviews}{}}}
\subsection{Method \code{showDatasetOverviews()}}{
Show overviews of purified analytical/metadata sets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderInputDataManager$showDatasetOverviews()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderInputDataManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderInputDataManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderInputDataManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
