% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBuilder.R
\name{ModelBuilderUtility}
\alias{ModelBuilderUtility}
\title{R6 class as utility of Model builer which is available for all the class, members and functions.}
\description{
R6 class as utility of Model builer which is available for all the class, members and functions.

R6 class as utility of Model builer which is available for all the class, members and functions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mlClassMethodEnum}}{We do not use this anymore !!!}

\item{\code{outputDir}}{We do not use this anymore !!!}

\item{\code{experimentId}}{String - experiment ID as first 12 characters of fullExperimentId.}

\item{\code{fullExperimentId}}{String - full experiment ID as SHA256 value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelBuilderUtility-new}{\code{ModelBuilderUtility$new()}}
\item \href{#method-ModelBuilderUtility-generateExperimentIds}{\code{ModelBuilderUtility$generateExperimentIds()}}
\item \href{#method-ModelBuilderUtility-initClassResultLog}{\code{ModelBuilderUtility$initClassResultLog()}}
\item \href{#method-ModelBuilderUtility-initClassResultStats}{\code{ModelBuilderUtility$initClassResultStats()}}
\item \href{#method-ModelBuilderUtility-getClassMlMethodArg}{\code{ModelBuilderUtility$getClassMlMethodArg()}}
\item \href{#method-ModelBuilderUtility-getRegMlMethodArg}{\code{ModelBuilderUtility$getRegMlMethodArg()}}
\item \href{#method-ModelBuilderUtility-getClassStats}{\code{ModelBuilderUtility$getClassStats()}}
\item \href{#method-ModelBuilderUtility-getRegStats}{\code{ModelBuilderUtility$getRegStats()}}
\item \href{#method-ModelBuilderUtility-errorBomb}{\code{ModelBuilderUtility$errorBomb()}}
\item \href{#method-ModelBuilderUtility-clone}{\code{ModelBuilderUtility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-new}{}}}
\subsection{Method \code{new()}}{
Inilialise ModelBuilderUtility - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-generateExperimentIds"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-generateExperimentIds}{}}}
\subsection{Method \code{generateExperimentIds()}}{
Generate SHA256 string which will be full experiment ID - it generates
full SHA256 hash (fullExperimentId) and its first 12 characters which
will be experiment ID (experimentId).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$generateExperimentIds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-initClassResultLog"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-initClassResultLog}{}}}
\subsection{Method \code{initClassResultLog()}}{
Initialise data frame of classification result log - we do not use this anymore !!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$initClassResultLog()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-initClassResultStats"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-initClassResultStats}{}}}
\subsection{Method \code{initClassResultStats()}}{
Initialise data frame of regression result log - we do not use this anymore !!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$initClassResultStats()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-getClassMlMethodArg"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-getClassMlMethodArg}{}}}
\subsection{Method \code{getClassMlMethodArg()}}{
Get classification ML method name as caret's argument - it is mainly
used for buildClassModel from MlModelBuilder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$getClassMlMethodArg(mlMethod)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mlMethod}}{String - ML method name (coming from buildClassModel).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-getRegMlMethodArg"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-getRegMlMethodArg}{}}}
\subsection{Method \code{getRegMlMethodArg()}}{
Get regression ML method name as caret's argument - it is mainly
used for buildRegModel from MlModelBuilder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$getRegMlMethodArg(mlMethod)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mlMethod}}{String - ML method name (coming from buildClassModel).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-getClassStats"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-getClassStats}{}}}
\subsection{Method \code{getClassStats()}}{
Get result statistics result table for classification by using caret's
confusionMatrix() function - it returns all the results of confusionMatrix()
function anyway. It is mainly used for buildClassModel from MlModelBuilder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$getClassStats(predictions, reals)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{Vector of character or string - predicted values.}

\item{\code{reals}}{Vector of character or string - real values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-getRegStats"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-getRegStats}{}}}
\subsection{Method \code{getRegStats()}}{
Get result statistics result table for regression: RMSE; R-Squared; MAE
and Accuracy - it is mainly used for buildRegModel from MlModelBuilder class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$getRegStats(predictions, reals)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{Vector of numeric - predicted values.}

\item{\code{reals}}{Vector of numeric - real values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-errorBomb"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-errorBomb}{}}}
\subsection{Method \code{errorBomb()}}{
BOOOOOOOOOM !!! !!! !!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$errorBomb(errorContent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorContent}}{String - detailed error message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelBuilderUtility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelBuilderUtility-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelBuilderUtility$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
