% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{modelFinder}
\alias{modelFinder}
\title{Search models by sensor name, product name and ML method name.}
\usage{
modelFinder(
  databaseName,
  sensorName = NULL,
  productName = NULL,
  mlName = NULL,
  modelId = NULL,
  type = "list"
)
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{sensorName}{String - a name of sensor data used for the prediction.}

\item{productName}{String - a name of product.}

\item{mlName}{String - a name of ML method.}

\item{modelId}{String - a model ID.}

\item{type}{String - data type of result experiment overviews. "list" or "json", default "list".}
}
\description{
Search models by sensor name, product name and ML method name.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
