% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{ModelViewerFunctionManager}
\alias{ModelViewerFunctionManager}
\title{R6 class which contains all the available functions for model viewer.}
\description{
R6 class which contains all the available functions for model viewer.

R6 class which contains all the available functions for model viewer.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{util}}{Class - class of ModelViewerUtility.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelViewerFunctionManager-new}{\code{ModelViewerFunctionManager$new()}}
\item \href{#method-ModelViewerFunctionManager-getProductsOverview}{\code{ModelViewerFunctionManager$getProductsOverview()}}
\item \href{#method-ModelViewerFunctionManager-getProductDetails}{\code{ModelViewerFunctionManager$getProductDetails()}}
\item \href{#method-ModelViewerFunctionManager-getMLmethodRankings}{\code{ModelViewerFunctionManager$getMLmethodRankings()}}
\item \href{#method-ModelViewerFunctionManager-detectSensors}{\code{ModelViewerFunctionManager$detectSensors()}}
\item \href{#method-ModelViewerFunctionManager-clone}{\code{ModelViewerFunctionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerFunctionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerFunctionManager-new}{}}}
\subsection{Method \code{new()}}{
Inilialise ModelViewerFunctionManager - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerFunctionManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerFunctionManager-getProductsOverview"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerFunctionManager-getProductsOverview}{}}}
\subsection{Method \code{getProductsOverview()}}{
Get products' overviews - it shows Title, Product, ModellingType,
AnalyticalData, MetaData and ID. The output results are in the list format
because it is used in the REST API functionality as well. This function is
used for productsOverseer() (exported) function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerFunctionManager$getProductsOverview(databaseName, filter = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{\code{filter}}{Object (list) - list of searching filter for available products.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerFunctionManager-getProductDetails"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerFunctionManager-getProductDetails}{}}}
\subsection{Method \code{getProductDetails()}}{
Get more detialed information of the product - e.g., 'Title', 'Description',
'Author', 'Product', 'Date', 'ModellingType' and so on.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerFunctionManager$getProductDetails(databaseName, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{\code{id}}{String - an ID which store models built by foodguardRanker::modelBuilder().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerFunctionManager-getMLmethodRankings"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerFunctionManager-getMLmethodRankings}{}}}
\subsection{Method \code{getMLmethodRankings()}}{
Get ML rankings of specific product - it reads and shows the content of
'databaseName/id/rankings'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerFunctionManager$getMLmethodRankings(databaseName, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{\code{id}}{String - an ID which store models built by foodguardRanker::modelBuilder().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerFunctionManager-detectSensors"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerFunctionManager-detectSensors}{}}}
\subsection{Method \code{detectSensors()}}{
Detect sensor name(s) by the input sensor ID - it reads
'databaseName/id/overview.json'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerFunctionManager$detectSensors(databaseName, sensorId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{\code{sensorId}}{String - an ID which store models built by foodguardRanker::modelBuilder().}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerFunctionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerFunctionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerFunctionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
