% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\title{Show overviews of available experiments and ther models.}
\usage{
predictor(
  databaseName,
  id,
  mlMethod,
  analyticalData,
  metadata,
  inputSensor,
  type = "list"
)
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{id}{String - an ID which store models built by foodguardRanker::modelBuilder().}

\item{mlMethod}{String - a name of ML method used for the prediction.}

\item{analyticalData}{String - a name of analytical data used for the prediction.}

\item{metadata}{String - a name of metadata used for the prediction.}

\item{inputSensor}{String - a filename of input sensor data used for the prediction.}

\item{type}{String - data type of result prediction. "list" or "json", default "list".}
}
\description{
Description here.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
