% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{PredictorFunctionManager}
\alias{PredictorFunctionManager}
\title{R6 class which contains all the available functions for prediction.}
\description{
R6 class which contains all the available functions for prediction.

R6 class which contains all the available functions for prediction.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{util}}{Class - class of PredictorUtility.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictorFunctionManager-new}{\code{PredictorFunctionManager$new()}}
\item \href{#method-PredictorFunctionManager-predictFreshness}{\code{PredictorFunctionManager$predictFreshness()}}
\item \href{#method-PredictorFunctionManager-clone}{\code{PredictorFunctionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorFunctionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorFunctionManager-new}{}}}
\subsection{Method \code{new()}}{
Inilialise PredictorUtility - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorFunctionManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorFunctionManager-predictFreshness"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorFunctionManager-predictFreshness}{}}}
\subsection{Method \code{predictFreshness()}}{
Description here.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorFunctionManager$predictFreshness(
  databaseName,
  id,
  mlMethod,
  analyticalData,
  metadata,
  inputSensorDf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{databaseName}}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{\code{id}}{String - an ID which store models built by foodguardRanker::modelBuilder().}

\item{\code{mlMethod}}{String - a name of ML method used for the prediction.}

\item{\code{analyticalData}}{String - a name of analytical data used for the prediction.}

\item{\code{metadata}}{String - a name of metadata used for the prediction.}

\item{\code{inputSensorDf}}{Object (data.frame) - a data frame of input sensor data used for the prediction.}

\item{\code{type}}{String - data type of result experiment overviews. "list" or "json", default "table".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorFunctionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorFunctionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorFunctionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
