% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{PredictorUtility}
\alias{PredictorUtility}
\title{R6 class as utility of Predictor which is available for all the class, members and functions.}
\description{
R6 class as utility of Predictor which is available for all the class, members and functions.

R6 class as utility of Predictor which is available for all the class, members and functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PredictorUtility-new}{\code{PredictorUtility$new()}}
\item \href{#method-PredictorUtility-getSensorId}{\code{PredictorUtility$getSensorId()}}
\item \href{#method-PredictorUtility-checkIfCsv}{\code{PredictorUtility$checkIfCsv()}}
\item \href{#method-PredictorUtility-checkIfJson}{\code{PredictorUtility$checkIfJson()}}
\item \href{#method-PredictorUtility-errorBomb}{\code{PredictorUtility$errorBomb()}}
\item \href{#method-PredictorUtility-clone}{\code{PredictorUtility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorUtility-new"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorUtility-new}{}}}
\subsection{Method \code{new()}}{
Inilialise PredictorUtility - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorUtility$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorUtility-getSensorId"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorUtility-getSensorId}{}}}
\subsection{Method \code{getSensorId()}}{
Create analytical data IDs - sensor's colnames are combined
and changed into SHA 256. Thise IDs are used for automated
prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorUtility$getSensorId(inputSensorDf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputSensorDf}}{Object - input sensor as data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorUtility-checkIfCsv"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorUtility-checkIfCsv}{}}}
\subsection{Method \code{checkIfCsv()}}{
Check if the input data is in CSV format - if CSV, return TRUE, otherwise FALSE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorUtility$checkIfCsv(inputSensor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputSensor}}{String - input sensor file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorUtility-checkIfJson"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorUtility-checkIfJson}{}}}
\subsection{Method \code{checkIfJson()}}{
Check if the input data is in JSON format - if JSON, return TRUE, otherwise FALSE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorUtility$checkIfJson(inputSensor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputSensor}}{String - input sensor file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorUtility-errorBomb"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorUtility-errorBomb}{}}}
\subsection{Method \code{errorBomb()}}{
BOOOOOOOOOM !!! !!! !!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorUtility$errorBomb(errorContent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorContent}}{String - detailed error message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PredictorUtility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PredictorUtility-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictorUtility$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
