% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{productDetailsViewer}
\alias{productDetailsViewer}
\title{Show detailed information of a specific product and their models.}
\usage{
productDetailsViewer(databaseName, id, type = "list")
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{id}{String - an ID which store models built by foodguardRanker::modelBuilder().}

\item{type}{String - data type of result experiment overviews. "list" or "json", default "list".}
}
\description{
It shows more detailed information of the product given by ID which are written in 'ID/overview.json'.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
