% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{productsOverseer}
\alias{productsOverseer}
\title{Show overviews of available experiments and their models.}
\usage{
productsOverseer(databaseName, type = "list")
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{type}{String - data type of result experiment overviews. "list" or "json", default "list".}
}
\description{
It shows overviews of available products and their models.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
