% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{sensorDetector}
\alias{sensorDetector}
\title{Detect types of input analytical datasets' name}
\usage{
sensorDetector(databaseName, inputSensor, type = "list")
}
\arguments{
\item{databaseName}{String - a name of database which store models built by foodguardRanker::modelBuilder().}

\item{type}{String - data type of result experiment overviews. "list" or "json", default "list".}

\item{sensorName}{String - a name of sensor data used for the prediction.}

\item{sensorFile}{String - a file path of sensor data used for the prediction.}

\item{productName}{String - a name of product.}
}
\description{
Detect types of input analytical datasets' name. Carry on grid-searching into the
database and find the name(s) the analytical data by SHA256.
}
\details{
This product was funded by the European Union project FoodGuard, grant number 101136542
}
