% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelViewer.R
\name{ModelViewerUtility}
\alias{ModelViewerUtility}
\title{R6 class as utility of Model viewer which is available for all the class, members and functions.}
\description{
R6 class as utility of Model viewer which is available for all the class, members and functions.

R6 class as utility of Model viewer which is available for all the class, members and functions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelViewerUtility-new}{\code{ModelViewerUtility$new()}}
\item \href{#method-ModelViewerUtility-getOverviewTable}{\code{ModelViewerUtility$getOverviewTable()}}
\item \href{#method-ModelViewerUtility-getProductDetailsTable}{\code{ModelViewerUtility$getProductDetailsTable()}}
\item \href{#method-ModelViewerUtility-getMLmethodRankingsTable}{\code{ModelViewerUtility$getMLmethodRankingsTable()}}
\item \href{#method-ModelViewerUtility-getSensorId}{\code{ModelViewerUtility$getSensorId()}}
\item \href{#method-ModelViewerUtility-checkIfCsv}{\code{ModelViewerUtility$checkIfCsv()}}
\item \href{#method-ModelViewerUtility-checkIfJson}{\code{ModelViewerUtility$checkIfJson()}}
\item \href{#method-ModelViewerUtility-errorBomb}{\code{ModelViewerUtility$errorBomb()}}
\item \href{#method-ModelViewerUtility-clone}{\code{ModelViewerUtility$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-new}{}}}
\subsection{Method \code{new()}}{
Inilialise ModelViewerUtility - so far it does nothing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-getOverviewTable"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-getOverviewTable}{}}}
\subsection{Method \code{getOverviewTable()}}{
Read and organise overview of models available for all the products in JSON format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$getOverviewTable(overviewJsonFiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{overviewJsonFiles}}{Vector of character - products overview JSON file paths.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-getProductDetailsTable"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-getProductDetailsTable}{}}}
\subsection{Method \code{getProductDetailsTable()}}{
Get detailed information of one product in table format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$getProductDetailsTable(productDirPath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{productDirPath}}{String - product dir path (database/ID).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-getMLmethodRankingsTable"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-getMLmethodRankingsTable}{}}}
\subsection{Method \code{getMLmethodRankingsTable()}}{
Get ML rankings of one experiment in table format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$getMLmethodRankingsTable(rankingsFilePath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rankingsFilePath}}{String - MODELS dir path (database/ID/rankings.csv).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-getSensorId"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-getSensorId}{}}}
\subsection{Method \code{getSensorId()}}{
Create analytical data IDs - sensor's colnames are combined
and changed into SHA 256. Thise IDs are used for automated
prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$getSensorId(inputSensorDf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputSensorDf}}{Object - input sensor as data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-checkIfCsv"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-checkIfCsv}{}}}
\subsection{Method \code{checkIfCsv()}}{
Check if the input data is in CSV format - if CSV, return TRUE, otherwise FALSE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$checkIfCsv(inputSensor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputSensor}}{String - input sensor file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-checkIfJson"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-checkIfJson}{}}}
\subsection{Method \code{checkIfJson()}}{
Check if the input data is in JSON format - if JSON, return TRUE, otherwise FALSE.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$checkIfJson(inputSensor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputSensor}}{String - input sensor file path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-errorBomb"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-errorBomb}{}}}
\subsection{Method \code{errorBomb()}}{
BOOOOOOOOOM !!! !!! !!!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$errorBomb(errorContent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{errorContent}}{String - detailed error message.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelViewerUtility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelViewerUtility-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelViewerUtility$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
